requirements.txt
Flask==2.2.2
Flask_Caching==2.0.1
geopy==2.2.0
numpy==1.23.3
Pillow==9.3.0
pymongo[srv]==4.2.0
python-dotenv==0.21.0
requests==2.25.1